+++ tests/test_config.py
@@ -4,7 +4,7 @@ from typing import Any
 import pytest
 import tomlkit as tm
 
-from lydian.config import Config, LogLevel, env_to_bool
+from lydian.config import Config, LogLevel
 from lydian.const import TESTS_DIR
 
 
@@ -24,10 +24,6 @@ def test_dump(tmpdir: Path) -> None:
     assert inst.vote_skipping.enabled == parsed['vote_skipping']['enabled']
     assert inst.logging.level == LogLevel(parsed['logging']['level'])
     assert inst.debug == parsed['debug']
-
-def test_env_to_bool() -> None:
-    assert env_to_bool('0') is env_to_bool('false') is env_to_bool('faLSE') is False
-    assert env_to_bool('1') is env_to_bool('true') is env_to_bool('trUE') is True
 
 def test_update_from_toml() -> None:
     inst = Config()