+++ tests/test_util_fromstr.py
@@ -3,6 +3,10 @@ import pytest
 from lydian.util import FromStr
 
 
+def test_to_bool() -> None:
+    assert FromStr.to_bool('0') is FromStr.to_bool('false') is FromStr.to_bool('faLSE') is False
+    assert FromStr.to_bool('1') is FromStr.to_bool('true') is FromStr.to_bool('trUE') is True
+
 @pytest.mark.parametrize(('s', 'expected'),
     [
         ('1 b',          1),